Additional minor simplifications and preparations for compatibility with JDK >= 8 (avoid deprecation warnings), plus a number of JavaDoc corrections

+++ demos/scalability/src/main/java/org/prevayler/demos/scalability/Main.java
@@ -337,7 +337,7 @@ public class Main {
 
   static private int intProperty(String name) {
     try {
-      return Integer.valueOf(property(name)).intValue();
+      return Integer.valueOf(property(name));
     } catch (NumberFormatException nfx) {
       out("NumberFormatException reading property " + name);
       throw nfx;
